Add broadcast and message handling extensions
